Maven-Published-Frameworks-1.0
While the publisher and the authors have used good faith efforts to ensure that the information and
instructions contained in this work are accurate, the publisher and the authors disclaim all responsibility
for errors or omissions, including without limitation responsibility for damages resulting from the use of
or reliance on this work. Use of the information and instructions contained in this work is at your own
risk. We supplement the core team with several "subteams". Each subteam is focused on a specific area, to ensure global coordination and a strong, coherent vision for the projects as a whole, each subteam is led by a member of the core team.